+++ packages/i18n/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @brickflow/i18n
 
+## 0.0.52
+
+### Patch Changes
+
+- Fix: remove auto import useTranslate
+
 ## 0.0.51
 
 ### Patch Changes